w07d5/act05 - Connect the favorite icon and modal
