+++ frontend/src/App.jsx
@@ -12,6 +12,24 @@ const App = () => {
   const [displayModal, setDisplayModal] = useState(false);
   const [selectedPhotoId, setSelectedPhotoId] = useState(null);
 
+  const [favourites, setFavourites] = useState([]);
+
+  const toggleFavourite = (photoId) => {
+    // If it does, it will remove it from the state
+    // OR
+    if (favourites.includes(photoId)) {
+      setFavourites((prev) => prev.filter((id) => id !== photoId));
+      // OR
+      // const copyOfFavourites = [...favourites].filter(
+      //   (favPhotoId) => favPhotoId !== photoId
+      // );
+      // setFavourites(copyOfFavourites);
+      return;
+    }
+    // Else it will add it if it doesn't exist
+    setFavourites((prev) => [...prev, photoId]);
+  };
+
   const viewModal = (photoId) => {
     setSelectedPhotoId(photoId);
     setDisplayModal(true);
@@ -23,6 +41,8 @@ const App = () => {
         photos={mockPhotoData}
         topics={mockTopicData}
         viewModal={viewModal}
+        favourites={favourites}
+        toggleFavourite={toggleFavourite}
       />
       {displayModal && (
         <PhotoDetailsModal
@@ -30,6 +50,8 @@ const App = () => {
           selectedPhoto={mockPhotoData.find(
             (photo) => photo.id === selectedPhotoId
           )}
+          toggleFavourite={toggleFavourite}
+          favourites={favourites}
         />
       )}
     </div>
